+++ lib/interact-for-command-params.ts
@@ -12,6 +12,7 @@ import { interactForAcsUser } from "./interact-for-acs-user"
 import { interactForCredentialPool } from "./interact-for-credential-pool"
 import { ContextHelpers } from "./types"
 import { interactForAcsEntrance } from "./interact-for-acs-entrance"
+import { interactForOpenApiObject } from "./interact-for-open-api-object"
 
 const ergonomicPropOrder = [
   "name",
@@ -42,249 +43,14 @@ export const interactForCommandParams = async (
 
   if (!("content" in requestBody)) throw new Error("No content in requestBody")
 
-  const schema: OpenApiSchema = flattenObjSchema(
-    (requestBody as any).content["application/json"].schema
-  )
+  const requestSchema = (requestBody as any).content["application/json"].schema
 
-  const { properties = {}, required = [] } = schema
-
-  const haveAllRequiredParams = required.every((k) => currentParams[k])
-
-  const propSortScore = (prop: string) => {
-    if (required.includes(prop)) return 100
-    if (currentParams[prop] !== undefined)
-      return 50 - Object.keys(currentParams).indexOf(prop)
-    return ergonomicPropOrder.indexOf(prop)
-  }
-
-  const cmdPath = `/${command.join("/").replace(/-/g, "_")}`
-  console.log("")
-  const { paramToEdit } = await prompts({
-    name: "paramToEdit",
-    message: `[${cmdPath}] Parameters`,
-    type: "autocomplete",
-    choices: [
-      ...(haveAllRequiredParams
-        ? [
-            {
-              value: "done",
-              title: `[Make API Call] ${cmdPath}`,
-            },
-          ]
-        : []),
-      ...Object.keys(properties)
-        .map((k) => {
-          let propDesc = (properties[k] as any)?.description ?? ""
-          return {
-            title: k + (required.includes(k) ? "*" : ""),
-            value: k,
-            description:
-              currentParams[k] !== undefined
-                ? `[${currentParams[k]}] ${propDesc}`
-                : propDesc,
-          }
-        })
-        .sort((a, b) => propSortScore(b.value) - propSortScore(a.value)),
-    ],
-  })
-
-  if (paramToEdit === "done") {
-    // TODO check for required
-    return currentParams
-  }
-
-  const prop = properties[paramToEdit]
-
-  if (paramToEdit === "device_id") {
-    const device_id = await interactForDevice()
-    return interactForCommandParams(
-      { command, params: { ...currentParams, device_id } },
-      ctx
-    )
-  } else if (paramToEdit === "access_code_id") {
-    const access_code_id = await interactForAccessCode(currentParams as any)
-    return interactForCommandParams(
-      {
-        command,
-        params: {
-          ...currentParams,
-          access_code_id,
-        },
-      },
-      ctx
-    )
-  } else if (paramToEdit === "connected_account_id") {
-    const connected_account_id = await interactForConnectedAccount()
-    return interactForCommandParams(
-      {
-        command,
-        params: {
-          ...currentParams,
-          connected_account_id,
-        },
-      },
-      ctx
-    )
-  } else if (paramToEdit === "user_identity_id") {
-    const user_identity_id = await interactForUserIdentity()
-    return interactForCommandParams(
-      {
-        command,
-        params: {
-          ...currentParams,
-          user_identity_id,
-        },
-      },
-      ctx
-    )
-  } else if (paramToEdit.endsWith("acs_system_id")) {
-    const acs_system_id = await interactForAcsSystem()
-    return interactForCommandParams(
-      {
-        command,
-        params: {
-          ...currentParams,
-          [paramToEdit]: acs_system_id,
-        },
-      },
-      ctx
-    )
-  } else if (paramToEdit.endsWith("credential_pool_id")) {
-    const credential_pool_id = await interactForCredentialPool()
-    return interactForCommandParams(
-      {
-        command,
-        params: {
-          ...currentParams,
-          [paramToEdit]: credential_pool_id,
-        },
-      },
-      ctx
-    )
-  } else if (paramToEdit.endsWith("acs_user_id")) {
-    const acs_user_id = await interactForAcsUser()
-    return interactForCommandParams(
-      {
-        command,
-        params: {
-          ...currentParams,
-          [paramToEdit]: acs_user_id,
-        },
-      },
-      ctx
-    )
-  } else if (paramToEdit.endsWith("acs_entrance_id")) {
-    const acs_entrance_id = await interactForAcsEntrance()
-    return interactForCommandParams(
-      {
-        command,
-        params: {
-          ...currentParams,
-          [paramToEdit]: acs_entrance_id,
-        },
-      },
-      ctx
-    )
-  } else if (
-    // TODO replace when openapi returns if a field is a timestamp
-    paramToEdit.endsWith("_at") ||
-    paramToEdit === "since" ||
-    paramToEdit.endsWith("_before") ||
-    paramToEdit.endsWith("_after")
-  ) {
-    const tsval = await interactForTimestamp()
-    return interactForCommandParams(
-      {
-        command,
-        params: {
-          ...currentParams,
-          [paramToEdit]: tsval,
-        },
-      },
-      ctx
-    )
-  }
-
-  if ("type" in prop) {
-    if (prop.type === "string") {
-      let value
-      if (prop.enum) {
-        value = (
-          await prompts({
-            name: "value",
-            message: `${paramToEdit}:`,
-            type: "select",
-            choices: prop.enum.map((v) => ({
-              title: v.toString(),
-              value: v.toString(),
-            })),
-          })
-        ).value
-      } else {
-        value = (
-          await prompts({
-            name: "value",
-            message: `${paramToEdit}:`,
-            type: "text",
-          })
-        ).value
-      }
-      return interactForCommandParams(
-        {
-          command,
-          params: {
-            ...currentParams,
-            [paramToEdit]: value,
-          },
-        },
-        ctx
-      )
-    } else if (prop.type === "boolean") {
-      const { value } = await prompts({
-        name: "value",
-        message: `${paramToEdit}:`,
-        type: "toggle",
-        initial: true,
-        active: "true",
-        inactive: "false",
-      })
-
-      return interactForCommandParams(
-        {
-          command,
-          params: {
-            ...currentParams,
-            [paramToEdit]: value,
-          },
-        },
-        ctx
-      )
-    } else if (prop.type === "array") {
-      const value = (
-        await prompts({
-          name: "value",
-          message: `${paramToEdit}:`,
-          type: "autocompleteMultiselect",
-          choices: (prop as any).items.enum.map((v: string) => ({
-            title: v.toString(),
-            value: v.toString(),
-          })),
-        })
-      ).value
-      return interactForCommandParams(
-        {
-          command,
-          params: {
-            ...currentParams,
-            [paramToEdit]: value,
-          },
-        },
-        ctx
-      )
-    }
-  }
-
-  throw new Error(
-    `Didn't know how to handle OpenAPI schema for property: "${paramToEdit}"`
+  return interactForOpenApiObject(
+    {
+      command: args.command,
+      params: args.params,
+      schema: requestSchema,
+    },
+    ctx
   )
 }
